patch for DateColumn::CUSTOM with Laravel datetime strings
Adding a simple patch to allow the use of DateColumn::CUSTOM with Laravel datetime columns in PostgreSQL, which return strings when accessed from the model object. It's really rough and sloppy, but I'm hoping this might start a discussion whereby a cleaner implementation is suggested. I don't mind reworking this to account for all the cases in DateColumn, and to make it a little more robust.