Add a new option, queryKeepsLimits defines if the LIMITs should stay
When exporting a query using the QueryEngine, by default the current
behaviour is to run the limit and skip functions on the QueryBuilder.

Sometimes this may be the behaviour that you want.

The default behaviour however should be to get the query without any
limit or skips. A new option is introduced to make sure that the limits
are the same as they were previously (where the query is the same as
what is shown in the datatable).

I will write a wiki article a little later showing a potential new use &
implementation of this new functionality, which will allow server-side
CSV exports ala TableTools (but for larger data).